+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/annotation/ControllerScanner.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.annotation;
 
 import com.interface21.context.stereotype.Controller;
 import org.reflections.Reflections;
