+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/annotation/ControllerScannerTest.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.annotation;
 
 import constructor.success.TestSuccessController;
 import org.junit.jupiter.api.Test;
